Allowed fetching tocProduct nodes for catalog

+++ CatalogDataTransformer/Models/TOC_Product.cs
@@ -4,5 +4,5 @@ public class TOCProduct
 {
     public int ProductId { get; set; }
     public int NodeId { get; set; }
-    public byte SeqNo { get; set; }
+    public short SeqNo { get; set; }
 }
