+++ KITSProductData/Data/Entities/TOCProduct.cs
@@ -1,4 +1,3 @@
-using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
 namespace Data.Entities;
@@ -8,7 +7,7 @@ public class TOCProduct
 {
     public int ProductId { get; set; }
     public int NodeId { get; set; } 
-    public byte SeqNo { get; set; }
+    public short SeqNo { get; set; }
     public CatalogTOC CatalogTOC { get; set; } = null!;
     public Product Product { get; set; } = null!;
 }
